[Spring 체스 - 2단계] 칙촉(안영윤) 미션 제출합니다. (#482)
* feat: 스파크 자바 체스게임 코드 가져온 후 spring mvc 적용

* refactor: 게임상태패턴에 인터페이스 적용 및 getBoard 의 반환값 Board 로 변경

- ready 의 응집도를 높임
- getBoard 의 반환값을 board 로 변경함으로써 변경에 유연하도록 함

* refator: spring jdbc 적용

* refactor: SimpleJdbcInsert 대신 JdbcTemplate 을 사용

* test: dao 테스트 코드 추가

* test: service 테스트 코드 추가

* test: controller 테스트 코드 추가

* test: controller 테스트 코드 추가

* feat: 추가 기능 구현 적용

* 충돌해결

* feat: 추가 기능 구현

* refactor: move 시 put 메서드 대신 patch 메서드 사용

* refactor: rest api 적용

* feat: service 클래스에 transactional 적용

* refactor: service 테스트시 사용하는 db h2로 변경

* refactor: restController, controller 분리

* refactor: ControllerAdvice 이름 변경

* style: 우테코 코드 컨벤션 적용

* refactor: ResponseEntity 생성시 정적 메서드 사용하도록 함

* refactor: ChessDao 테스트 코드 리펙토링

* refactor: ChessDao 테스트 코드 리펙토링

* refactor: NPE 대신 IllegalArgumentException throw 해줌

* docs: gitignore 에서 init.sql 파일 제외

* feat: board 값이 비어있는지 확인하는 기능 구현

* feat: 모든 예외를 핸들링 하도록 구현

* feat: 비밀번호 불일치 시 예외발생 기능구현

* feat: board 로 게임 저장 및 게임 턴 업테이트 기능 추가

* feat: 정적 팩토리 메서드 추가

* refactor: transactional 메서드에 적용 및 코드 리펙토링

* docs: init.sql 파일 추가

* refactor: 메서드 이름 변경

* refactor: Dao 적용 어노테이션 변경

* refactor: 중복 코드 메서드로 분리

* refactor: 예상하지 못한 예외에 대해서 최소한의 정보만 반환하도록 리펙토링

* refactor: 예상하지 못한 예외에 대해서 최소한의 정보만 반환하도록 리펙토링

* refactor: 말 이동시 실행 로직 변경

- 전체삭제 대신 해당하는 값만 지우도록 리펙토링함

* refactor: 사용하지 않는 메서드 삭제

+++ src/main/java/chess/service/ChessService.java
@@ -8,18 +8,16 @@ import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceFactory;
 import chess.domain.position.Position;
-import chess.domain.state.Ready;
-import chess.domain.state.Running;
 import chess.domain.state.StateFactory;
 import chess.dto.ChessGameDto;
+import chess.dto.GameRoomDto;
+import chess.dto.MoveDto;
 import chess.dto.PieceAndPositionDto;
-import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-
-import static chess.dao.ChessDaoImpl.DEFAULT_GAME_ID;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class ChessService {
@@ -27,53 +25,80 @@ public class ChessService {
     private static final String FIRST_COLOR = "WHITE";
 
     private final ChessDao chessDao;
-    private ChessGame chessGame = new ChessGame(new Ready());
 
     public ChessService(final ChessDao chessDao) {
         this.chessDao = chessDao;
     }
 
-    public ChessGameDto newGame() {
-        deleteOldData(DEFAULT_GAME_ID);
-        initNewChessGame(DEFAULT_GAME_ID);
-        chessGame = new ChessGame(new Running(Color.from(chessDao.findCurrentColor(DEFAULT_GAME_ID)), convertToBoard(chessDao.findAllPiece(DEFAULT_GAME_ID))));
-        return new ChessGameDto(chessDao.findAllPiece(DEFAULT_GAME_ID), chessGame.status());
+    public int createRoom(final GameRoomDto gameRoomDto) {
+        return chessDao.initGame(gameRoomDto.getTitle(), gameRoomDto.getPassword());
     }
 
-    private void deleteOldData(final int gameId) {
+    @Transactional
+    public ChessGameDto resetGame(final int gameId) {
         chessDao.deleteAllPiece(gameId);
-        chessDao.deleteGame(gameId);
+        chessDao.saveGame(gameId, new Board(new BoardInitializer()));
+        chessDao.updateTurn(FIRST_COLOR, gameId);
+
+        return ChessGameDto.from(findChessGameById(gameId));
     }
 
-    private void initNewChessGame(final int gameId) {
-        chessDao.initGame(gameId);
-        chessDao.updateTurn(FIRST_COLOR, gameId);
-        for (final Map.Entry<Position, Piece> entry : new BoardInitializer().init().entrySet()) {
-            chessDao.savePiece(gameId, new PieceAndPositionDto(entry.getKey(), entry.getValue()));
-        }
+    public List<GameRoomDto> findAllGame() {
+        return chessDao.findAllGame();
+    }
+
+    public ChessGame findChessGameById(int gameId) {
+        return new ChessGame(
+                StateFactory.of(Color.from(chessDao.findCurrentColor(gameId)),
+                        convertToBoard(chessDao.findAllPiece(gameId)))
+        );
     }
 
     private Board convertToBoard(final List<PieceAndPositionDto> pieceAndPositionDtos) {
         final Map<Position, Piece> board = pieceAndPositionDtos.stream()
-                .collect(Collectors.toMap(it -> Position.from(it.getPosition()), it -> PieceFactory.of(it.getPieceName(), it.getPieceColor())));
+                .collect(Collectors.toMap(
+                        it -> Position.from(it.getPosition()),
+                        it -> PieceFactory.of(it.getPieceName(), it.getPieceColor()))
+                );
         return new Board(() -> board);
     }
 
-    public ChessGameDto move(final String from, final String to) {
+    @Transactional
+    public ChessGameDto move(final MoveDto moveDto) {
+        final var gameId = moveDto.getGameId();
+        final var chessGame = findChessGameById(gameId);
+        final var from = moveDto.getFrom();
+        final var to = moveDto.getTo();
+
         chessGame.move(Position.from(from), Position.from(to));
-        final var nextColor = Color.from(chessDao.findCurrentColor(DEFAULT_GAME_ID)).next();
-        updateBoard(from, to, DEFAULT_GAME_ID, nextColor.name());
-        return new ChessGameDto(chessDao.findAllPiece(DEFAULT_GAME_ID), chessGame.status());
+
+        updatePieces(gameId, from, to);
+
+        var currentColor = chessDao.findCurrentColor(gameId);
+        var nextColor = Color.from(currentColor).next().toString();
+        chessDao.updateTurn(nextColor, gameId);
+
+        return ChessGameDto.from(findChessGameById(gameId));
     }
 
-    private void updateBoard(final String from, final String to, final int gameId, final String color) {
+    private void updatePieces(int gameId, String from, String to) {
         chessDao.deletePiece(gameId, to);
         chessDao.updatePiece(from, to, gameId);
-        chessDao.updateTurn(color, gameId);
     }
 
-    public ChessGameDto loadGame() {
-        chessGame = new ChessGame(StateFactory.of(Color.from(chessDao.findCurrentColor(DEFAULT_GAME_ID)), convertToBoard(chessDao.findAllPiece(DEFAULT_GAME_ID))));
-        return new ChessGameDto(chessDao.findAllPiece(DEFAULT_GAME_ID), chessGame.status());
+    @Transactional
+    public void deleteGame(final GameRoomDto gameRoomDto) {
+        final var gameId = gameRoomDto.getGameId();
+        final var chessGame = findChessGameById(gameId);
+        final var room = chessDao.findRoomById(gameId);
+
+        room.checkPassword(gameRoomDto.getPassword());
+
+        if (chessGame.isRunning()) {
+            throw new IllegalArgumentException("게임이 종료되지 않았습니다.");
+        }
+
+        chessDao.deleteAllPiece(gameId);
+        chessDao.deleteGame(gameId);
     }
 }
